Rework the core data operation model as {Model}, {Model}+{Config}={Spec}
This commit brings in a new data model which cleans up the myriad of
diff classes that exist in the current Loom distribution.

Operators define models. End users send Configs and Loom merges the two
together into a Spec which it can then convert into Kops, Terraform or
whatever else.